add safeguard to not report errors when dir empty so terraform happy even when doing nothing

+++ terraform/build-idx-scripts.sh
@@ -9,10 +9,12 @@ fi
 
 instancetype=$1
 
-
+mkdir -p ./build-idx
 find ./build-idx -type f -exec rm {} \;
 mkdir -p ./build-idx/usr/local/bin/ ./build-idx/opt/splunk/scripts
-cp -p ./scripts-template/splunkconf-aws-terminate-idx.sh ./build-idx/usr/local/bin/
-chmod 555 ./build-idx/usr/local/bin/splunkconf-aws-terminate-idx.sh
+if [ -e "./scripts-template/splunkconf-aws-terminate-idx.sh" ]; then
+  cp -p ./scripts-template/splunkconf-aws-terminate-idx.sh ./build-idx/usr/local/bin/
+  chmod 555 ./build-idx/usr/local/bin/splunkconf-aws-terminate-idx.sh
+fi
 mkdir -p ../buckets/bucket-backup/splunkconf-backup/${instancetype}
 tar -C"./build-idx" --exclude "._*" --exclude ".DS*" -zcvf ../buckets/bucket-backup/splunkconf-backup/${instancetype}/backupconfsplunk-scripts-initial.tar.gz ./opt/splunk/scripts ./usr/local/bin
