+++ terraform/build-nonidx-scripts.sh
@@ -9,10 +9,12 @@ fi
 
 instancetype=$1
 
-
+mkdir -p ./build-nonidx
 find ./build-nonidx -type f -exec rm {} \;
 mkdir -p ./build-nonidx/usr/local/bin/ ./build-nonidx/opt/splunk/scripts
-cp -p ./scripts-template/aws-update-dns.sh ./build-nonidx/usr/local/bin/aws-postinstall.sh
-chmod 555 ./build-idx/usr/local/bin/aws-postinstall.sh
+if [ -e "./scripts-template/aws-update-dns.sh" ]; then
+  cp -p ./scripts-template/aws-update-dns.sh ./build-nonidx/usr/local/bin/aws-postinstall.sh
+  chmod 555 ./build-idx/usr/local/bin/aws-postinstall.sh
+fi
 mkdir -p ../buckets/bucket-backup/splunkconf-backup/${instancetype}
 tar -C"./build-idx" --exclude "._*" --exclude ".DS*" -zcvf ../buckets/bucket-backup/splunkconf-backup/${instancetype}/backupconfsplunk-scripts-initial.tar.gz ./opt/splunk/scripts ./usr/local/bin
